# sdk/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "whiskd daemon API",
    "version": "17",
    "description": "HTTP/JSON API exposed by the whiskd daemon on loopback. Generated from Go structs; do not edit by hand."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ],
  "paths": {
...
    "/v1/artifacts": {
      "get": {
        "operationId": "listArtifacts",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "workItemId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Artifact"
                  },
                  "type": "array"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
...
    "/v1/gate-reports": {
      "get": {
        "operationId": "listGateReports",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "workItemId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GateReport"
                  },
                  "type": "array"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/gate-reports/{gateReportID}/complete": {
      "post": {
        "operationId": "completeGate",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "gateReportID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteGateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateReport"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
...
    "/v1/project-attachments/{attachmentID}/delete": {
      "post": {
        "operationId": "deleteProjectAttachment",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "attachmentID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteProjectAttachmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/project-attachments/{attachmentID}/update": {
      "post": {
        "operationId": "updateProjectAttachment",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "attachmentID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectAttachmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "tags": [
          "workitems"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      },
      "post": {
        "operationId": "createProject",
        "tags": [
          "workitems"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/projects/{projectID}/attachments": {
      "post": {
        "operationId": "addProjectAttachment",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "projectID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProjectAttachmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/projects/{projectID}/context": {
      "get": {
        "operationId": "getProjectContext",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "projectID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectContext"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/projects/{projectID}/detail": {
      "get": {
        "operationId": "getProjectDetail",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "projectID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/projects/{projectID}/update": {
      "post": {
        "operationId": "updateProject",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "projectID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/prompt-templates": {
      "get": {
        "operationId": "listPromptTemplates",
        "tags": [
          "workitems"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PromptTemplate"
                  },
                  "type": "array"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
...
    "/v1/questions/{questionID}/answer": {
      "post": {
        "operationId": "answerQuestion",
        "tags": [
          "workitems"
        ],
        "parameters": [
          {
            "name": "questionID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerQuestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
...
    "/v1/workflow-templates": {
      "get": {
        "operationId": "listWorkflowTemplates",
        "tags": [
          "workitems"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowTemplate"
                  },
                  "type": "array"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Artifact": {
        "properties": {
          "body": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MetadataValue"
            },
            "type": "object"
          },
          "projectId": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "workItemId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "workItemId",
          "kind",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    }
  }
}